+++ conpot/protocols/http/command_responder.py
@@ -28,8 +28,8 @@ import http.server
 import http.client
 import os
 from lxml import etree
-from conpot.helpers import str_to_bytes
 import conpot.core as conpot_core
+from conpot.utils.networking import str_to_bytes
 import gevent
