+++ conpot/protocols/kamstrup_management/kamstrup_management_server.py
@@ -20,8 +20,8 @@ import gevent
 from gevent.server import StreamServer
 import conpot.core as conpot_core
 from .command_responder import CommandResponder
-from conpot.helpers import str_to_bytes
 from conpot.core.protocol_wrapper import conpot_protocol
+from conpot.utils.networking import str_to_bytes
 
 logger = logging.getLogger(__name__)
