+++ conpot/protocols/kamstrup_meter/decoder_382.py
@@ -16,10 +16,11 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 import logging
-from conpot.helpers import chr_py3
-from crc16.crc16pure import crc16xmodem
-from . import kamstrup_constants
 
+from crc16.crc16pure import crc16xmodem
+
+from conpot.utils.networking import chr_py3
+from . import kamstrup_constants
 
 logger = logging.getLogger(__name__)
