+++ conpot/protocols/kamstrup_meter/kamstrup_server.py
@@ -21,8 +21,8 @@ import random
 import socket
 from gevent.server import StreamServer
 import gevent
-from conpot.helpers import chr_py3
 import conpot.core as conpot_core
+from conpot.utils.networking import chr_py3
 from .request_parser import KamstrupRequestParser
 from .command_responder import CommandResponder
 from conpot.core.protocol_wrapper import conpot_protocol
