+++ conpot/protocols/s7comm/cotp.py
@@ -4,8 +4,8 @@
 
 from struct import pack, unpack
 import struct
-from conpot.helpers import str_to_bytes
 from conpot.protocols.s7comm.exceptions import ParseException
+from conpot.utils.networking import str_to_bytes
 
 
 class COTP(object):
