+++ conpot/protocols/tftp/tftp_server.py
@@ -25,9 +25,9 @@ from conpot.protocols.tftp import tftp_handler
 from gevent.server import DatagramServer
 import conpot.core as conpot_core
 from conpot.core.protocol_wrapper import conpot_protocol
+from conpot.utils.networking import get_interface_ip
 from tftpy import TftpException, TftpTimeout
 import logging
-from conpot.utils.ext_ip import get_interface_ip
 
 logger = logging.getLogger(__name__)
