+++ conpot/tests/test_kamstrup_meter.py
@@ -22,8 +22,8 @@ import conpot
 from conpot.protocols.kamstrup_meter.command_responder import CommandResponder
 from conpot.protocols.kamstrup_meter.request_parser import KamstrupRequestParser
 from conpot.protocols.kamstrup_meter.kamstrup_server import KamstrupServer
-from conpot.helpers import chr_py3
 from conpot.utils.greenlet import spawn_test_server, teardown_test_server
+from conpot.utils.networking import chr_py3
 from gevent import socket
 import os
 import unittest
