+++ docs/source/api/reference/conpot.utils.rst
@@ -20,6 +20,14 @@ conpot.utils.mac\_addr module
    :undoc-members:
    :show-inheritance:
 
+conpot.utils.networking module
+------------------------------
+
+.. automodule:: conpot.utils.networking
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 
 Module contents
 ---------------
